+++ dev-packages/e2e-tests/test-applications/nextjs-13/app/rsc-error/page.tsx
@@ -1,0 +1,6 @@
+export const dynamic = 'force-dynamic';
+
+export default async function Page() {
+  throw new Error('RSC error');
+  return <p>Hello World</p>;
+}