+++ packages/nextjs/test/integration/pages/api/http/index.ts
@@ -1,5 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-
-export default (_req: NextApiRequest, res: NextApiResponse) => {
-  res.status(200).json({});
-};